requirements.txt
openai-whisper 
gtts 
gradio 
groq